Add files via upload
Develop a java application with Employee class with Emp_name, Emp_id, Address, 
Mail_id, Mobile_no as members. Inherit the classes, Programmer, Assistant 
Professor, Associate Professor and Professor from employee class. Add Basic Pay 
(BP) as the member of all the 74 inherited classes with 97% of BP as DA, 10 % of BP 
as HRA, 12% of BP as PF, 0.1% of BP for staff club fund. Generate pay slips for the 
employees with their gross and net salary.